requirements.txt
catboost==0.21
dill==0.3.1.1
Flask==1.1.1
pandas==0.25.1
scikit-learn==0.22.2.post1
sklearn-pandas==1.8.0
gunicorn
Jinja2==2.10.3
